revert redundant changes
