+++ core/src/main/scala/org/ensime/core/SymbolToFqn.scala
@@ -23,7 +23,22 @@ import org.slf4j.Logger
  */
 trait SymbolToFqn { self: Global with PresentationCompilerBackCompat =>
   def logger: Logger
-  import FqnUtils._
+
+  import ClassName._
+  val ScalaPackageName: PackageName = PackageName(List("scala"))
+  val normaliseClass: ClassName => ClassName = Map(
+    ClassName(PackageName(List("scala", "runtime")), "BoxedUnit") -> PrimitiveVoid,
+    ClassName(ScalaPackageName, "<byname>") -> ClassName(ScalaPackageName, "Function0"),
+    ClassName(ScalaPackageName, "Boolean") -> PrimitiveBoolean,
+    ClassName(ScalaPackageName, "Byte") -> PrimitiveByte,
+    ClassName(ScalaPackageName, "Char") -> PrimitiveChar,
+    ClassName(ScalaPackageName, "Short") -> PrimitiveShort,
+    ClassName(ScalaPackageName, "Int") -> PrimitiveInt,
+    ClassName(ScalaPackageName, "Long") -> PrimitiveLong,
+    ClassName(ScalaPackageName, "Float") -> PrimitiveFloat,
+    ClassName(ScalaPackageName, "Double") -> PrimitiveDouble,
+    ClassName(ScalaPackageName, "Void") -> PrimitiveVoid
+  ).withDefault(identity)
 
   private def packageName(sym: Symbol): PackageName = {
     PackageName(sym.ownerChain.takeWhile(!_.isRootSymbol).reverse.map(_.encodedName))
